+++ device/VoltageCalculator.h
@@ -5,13 +5,13 @@
 class VoltageCalculator
 {
   private:
-    unsigned long _operatingVoltage;
+    float _operatingVoltage;
     int _resolution;
     float CalculateFinalResolution(int bits);
   public:
     VoltageCalculator();
-    VoltageCalculator(unsigned long operatingVoltage);
-    unsigned long GetOperatingVoltage();
+    VoltageCalculator(float operatingVoltage);
+    float GetOperatingVoltage();
     int GetResolution();
     float CalculateVoltage(int analogValue);
 };
